+++ src/cartservice/src/services/CartService.cs
@@ -3,7 +3,9 @@
 using System.Diagnostics;
 using System.Threading.Tasks;
 using Grpc.Core;
+using OpenTelemetry.Trace;
 using cartservice.cartstore;
+using cartservice.featureflags;
 using Oteldemo;
 
 namespace cartservice.services
@@ -11,11 +13,14 @@ namespace cartservice.services
     public class CartService : Oteldemo.CartService.CartServiceBase
     {
         private readonly static Empty Empty = new Empty();
+        private readonly static ICartStore BadCartStore = new RedisCartStore("badhost:1234");
         private readonly ICartStore _cartStore;
+        private readonly FeatureFlagHelper _featureFlagHelper;
 
-        public CartService(ICartStore cartStore)
+        public CartService(ICartStore cartStore, FeatureFlagHelper featureFlagService)
         {
             _cartStore = cartStore;
+            _featureFlagHelper = featureFlagService;
         }
 
         public async override Task<Empty> AddItem(AddItemRequest request, ServerCallContext context)
@@ -52,7 +57,24 @@ namespace cartservice.services
             activity?.SetTag("app.user.id", request.UserId);
             activity?.AddEvent(new("Empty cart"));
 
-            await _cartStore.EmptyCartAsync(request.UserId);
+            try
+            {
+                if (await _featureFlagHelper.GenerateCartError())
+                {
+                    await BadCartStore.EmptyCartAsync(request.UserId);
+                }
+                else
+                {
+                    await _cartStore.EmptyCartAsync(request.UserId);
+                }
+            }
+            catch (RpcException ex)
+            {
+                Activity.Current?.RecordException(ex);
+                Activity.Current?.SetStatus(ActivityStatusCode.Error, ex.Message);
+                throw;
+            }
+
             return Empty;
         }
     }